Made project run with debugger in VsCode. No using ./makefile, but by pressing F5.

+++ main.h
@@ -2,8 +2,8 @@
 #define MAIN_H
 
 // Global screen size
-#define SCREEN_WIDTH GetScreenWidth()
-#define SCREEN_HEIGHT GetScreenHeight()
+#define SCREEN_WIDTH (float)GetScreenWidth()
+#define SCREEN_HEIGHT (float)GetScreenHeight()
 
 #include "raylib.h"
